datamodel: add backhaul SSID on (local) alDevice
Although not explicitly mentioned in the Multi-AP specification, the
assumption is that there is at most one backhaul SSID configured on a
device. So keep track of this in the alDevice.

Add localDeviceUpdateBackhaulSsid() to set the backhaul SSID, and call
this from wscProcessM2() when a backhaul SSID is configured.

Add a setBackhaulSsid to struct radio. There is a separate callback per
radio to allow different implementations for different radios.