datamodel: keep track of configured radios
Add a member "configured" to the radio data structure to keep track if
it has been configured or not.

Rewrite _triggerAPSearchProcess() to use this state rather than
iterating over all interfaces.

Send out the AP Autoconfiguration Search message on all interfaces
rather than just the unconfigured one (which wouldn't have worked
anyway, since it's unconfigured...). Create a new function
send1905Multicast() to handle this case instead of send1905RawPacket().
Note that it should only be sent out on secured interfaces, which is
not yet handled in this function.

Also keep a global "configured" state which is used to initialise the
state of each radio.