UCI: work around broken string handling in UCI
Apparently, UCI assumes that the strings it gets from ubus are
0-terminated. This is broken because SSIDs and keys may contain embedded
0 characters (admittedly, not many devices would be able to deal with
that anyway). However, even for strings without embedded 0 characters,
it seems that UCI ignores the length of the string and just assumes it
is 0-terminated.

Therefore, when adding SSID and key strings, first copy into a buffer
where we can add a 0 terminator.