UCI: fix parsing of uci get prplmesh
The result of uci get is wrapped in a "values" field, and additionally
wrapped with the section name.

Since the section name is known to be "prplmesh", just add it as part of
the uci get.

Parse the "values" with an additional policy.

Also, we were using the "port" value instead of "al_address".