Use interface instead of address in receive path
The fake interfaces that are added for backhaul STAs all have the same
MAC address (the BSSID). Therefore, the MAC address is not a suitable
identifier for the interface on which the packet was received.

Therefore, pass around a struct interface instead of a MAC address.

Since this contains all needed information, there is no longer a need
for struct event1905Packet.

Note that the interface name is a suitable identifier (the name is
different) and that is what is used in the send path. So the al_send
functions don't need to be changed.