al_entity: automatically do AP autoconfiguration
AP autoconfiguration is supposed to be triggered by push button
configuration. However, when we do backhaul onboarding, the interface
may not be fully up yet by the time we trigger the prplmesh push button
procedure. In particular, the receive loop on that interface probably
errored out before so we have to wait until it is restarted.

Therefore, we simply repeat the AP autoconfiguration process
indefinitely - but only if we have a backhaul link.